Introduce own processor type for Ski
src/state.c: Configure own processor type for Ski

So far this was configured as McKinley, but the exposed processor
features (branchlong, 16-byte atomic ops) are actually on par with a
Montecito (family 32) and exceed even those of a Madison (branchlong)
which is a family 31/model 2 processor. Hence the naming in recent Linux
kernels with Ski/HP-Sim support enabled:

MonteSkito for a family 31/model 3 processor.

Signed-off-by: Johnny Mnemonic <jm@machine-hall.org>