Make tap networking work beyond Ski's host system
So far communication only worked between host and ski instance with
tun/tap networking. This because the MAC address of the tap interface
on the host was also used identically for the emulated network interface
in the Ski instance. Change the first two bytes to something familiar to
correct this and that way enable communications beyond the host system.

Suggested-by: naguam <naguam@cloudbsd.xyz>
Signed-off-by: Johnny Mnemonic <jm@machine-hall.org>