ski-bootloader/ski-efi: enable host idling via SSC_HALT_LIGHT
When the ia64 kernel idles it calls PAL_HALT_LIGHT,
but that is a no-op when running on ski.  This prevents
ski and the host from idling.

Update pal_emulator_static to recognize PAL_HALT_LIGHT
and to propagate it to ski as a SSC_HALT_LIGHT.  To have
any effect this requires a corresponding ski update to
implement SSC_HALT_LIGHT.

Tested with ski-1.3.2 + ITC 2MHz + SSC_HALT_LIGHT patches
on x86_64.

Signed-off-by: Mikael Pettersson <mikpelinux@gmail.com>
Signed-off-by: Johnny Mnemonic <jm@machine-hall.org>